refactor: AuthAcceptanceTest의 Support를 추가해서 전역으로 사용할 메서드 분리
