+++ src/test/java/nextstep/subway/auth/acceptance/AuthAcceptanceTest.java
@@ -1,25 +1,15 @@
 package nextstep.subway.auth.acceptance;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.AcceptanceTest;
-import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.member.acceptance.MemberAcceptanceTestSupport;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import static nextstep.subway.member.acceptance.MemberAcceptanceTestSupport.*;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
-public class AuthAcceptanceTest extends AcceptanceTest {
+class AuthAcceptanceTest extends AuthAcceptanceTestSupport {
 
     @BeforeEach
     @Override
@@ -55,36 +45,5 @@ public class AuthAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> findResponse = 내_정보_조회_요청(invalidAccessToken);
         // then
         내_정보_조회_실패(findResponse);
-    }
-
-    public static ExtractableResponse<Response> 토큰_발급_요청(String email, String password) {
-        Map<String, String> params = new HashMap<>();
-        params.put("email", email);
-        params.put("password", password);
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when().post("/login/token")
-                .then().log().all().extract();
-    }
-
-    public static String getAccessToken(ExtractableResponse<Response> response) {
-        return response.as(TokenResponse.class).getAccessToken();
-    }
-
-    public static void 토큰_생성_완료(ExtractableResponse<Response> response) {
-        assertAll(
-                () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
-                () -> assertThat(response.as(TokenResponse.class).getAccessToken()).isNotEmpty()
-        );
-    }
-
-    private void 토큰_발급_실패(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
-    }
-
-    private void 내_정보_조회_실패(ExtractableResponse<Response> findResponse) {
-        assertThat(findResponse.statusCode()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
     }
 }
